requirements.txt
python==3.7.10
tensorflow==2.5.0
flask==2.0.1
cudatoolkit==11.2.0
numpy==1.19.5
pillow==8.3.1
requests==2.26.0